refactor: getInitialMusicDate를 피처 레이어로 이동 및 KST 보정 적용
